+++ completions/fancy-tar.zsh
@@ -1,18 +1,19 @@
-#compdef fancy-tar
+#compdef fancy-tar fancytar ftar
 
-_arguments \
-  '-o+[Output filename]:file:_files' \
-  '-n[No compression]' \
-  '-s[Slow mode]' \
-  '-x[Open output folder]' \
-  '-t[Tree preview]' \
-  '--tree[Tree preview]' \
-  '--no-recursion[Do not include subfolders]' \
-  '--hash[Generate SHA256 checksum]' \
-  '--encrypt=[Encryption method]:method:(gpg openssl)' \
-  '--recipient[Recipient GPG key]' \
-  '--password[Password for encryption]' \
-  '--zip[Create ZIP archive]' \
-  '--version[Show version]' \
-  '-h[Show help]' \
-  '--help[Show help]'
+_arguments -s \
+  '-o[Specify output file name]:output file:_files' \
+  '-n[Create uncompressed tar archive]' \
+  '-s[Use slower but better compression]' \
+  '-x[Open the output folder when done]' \
+  '-t[Show hierarchical file structure before archiving]' \
+  '--tree[Show hierarchical file structure before archiving]' \
+  '--no-recursion[Do not include directory contents]' \
+  '--hash[Output SHA256 hash file alongside the archive]' \
+  '--encrypt=[Encrypt archive with gpg or openssl]:method:(gpg openssl)' \
+  '--recipient=[Recipient ID for GPG public key encryption]:recipient:' \
+  '--password=[Password to use for encryption]:password:' \
+  '--zip[Create a .zip archive]' \
+  '--print-filename[Output only the final archive filename]' \
+  '--version[Show version information]' \
+  '--help[Show help message]' \
+  '*:files:_files'