+++ scripts/fancy_tar_progress.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-VERSION="1.7.2"
+VERSION="1.7.3"
 
 # Function to get next available filename
 get_next_filename() {